Update Exhibition page and add new components
