--- src/pages/Exhibition/Exhibition.test.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import styles from './Exhibition.module.css';
-
-function Exhibition() {
-  return (
-    <div className={styles.container}>
-      <h1 className={styles.title}>Welcome to the Home Page</h1>
-      <p>This is the main page of the application.</p>
-    </div>
-  );
-}
-
-export default Exhibition;
